update content structure
