+++ src/pages/geral/index.tsx
@@ -6,7 +6,7 @@ import fs from 'fs';
 import Footer from '../../components/Footer';
 import Navbar from '../../components/Navbar';
 
-const Programacao = ({ data, title }) => {
+const Reflexoes = ({ data, title }) => {
   const RealData = data.map((blog) => matter(blog));
   let ListItems = RealData.map((listItem) => listItem.data);
   ListItems = ListItems.sort((a, b) => {
@@ -23,11 +23,11 @@ const Programacao = ({ data, title }) => {
         <meta charSet="utf-8" />
         <title>{title}</title>
       </Head>
-      <Navbar hrefReturn="/" articleMetaData={{ title: 'Programação' }} />
+      <Navbar hrefReturn="/" articleMetaData={{ title: 'Reflexões' }} />
       <div className="flex flex-col justify-between py-16">
         <div className="py-5 px-2">
           <h1 className="text-center text-2xl font-bold text-accent-1 mb-6 md:text-4xl lg:text-5xl">
-            Programação
+            Reflexões
           </h1>
           {ListItems.map((blog, i) => (
             <div
@@ -37,7 +37,7 @@ const Programacao = ({ data, title }) => {
               <div
                 className="w-full md:w-1/4 h-40 rounded md:rounded"
                 style={{
-                  backgroundImage: `url('/img/content/programacao/${blog.slug}/thumb.png')`,
+                  backgroundImage: `url('/images/image.png')`,
                   backgroundSize: 'cover'
                 }}
               ></div>
@@ -46,7 +46,7 @@ const Programacao = ({ data, title }) => {
               <div className="md:w-3/5 p-4 md:h-40">
                 <div className="mb-8">
                   <div className="text-accent-1 font-bold text-xl mb-2">
-                    <Link href={`/programacao/${blog.slug}`}>{blog.title}</Link>
+                    <Link href={`/reflexoes/${blog.slug}`}>{blog.title}</Link>
                   </div>
                   <p className="text-accent-1-darker text-base">
                     {blog.description}
@@ -55,7 +55,7 @@ const Programacao = ({ data, title }) => {
                 <div className="flex items-center">
                   <img
                     className="w-10 h-10 rounded-full mr-4"
-                    src={`/img/avatar/${blog.avatar}`}
+                    src="/avatar/profile.jpg"
                     alt="Avatar of Tom Benevides"
                   />
                   <div className="text-sm">
@@ -75,14 +75,14 @@ const Programacao = ({ data, title }) => {
   );
 };
 
-export default Programacao;
+export default Reflexoes;
 
 export async function getStaticProps() {
-  const files = fs.readdirSync(`${process.cwd()}/content/programacao`, 'utf-8');
+  const files = fs.readdirSync(`${process.cwd()}/content/geral`, 'utf-8');
 
   const blogs = files.filter((fn) => fn.endsWith('.md'));
   const data = blogs.map((blog) => {
-    const path = `${process.cwd()}/content/programacao/${blog}`;
+    const path = `${process.cwd()}/content/geral/${blog}`;
     const rawContent = fs.readFileSync(path, {
       encoding: 'utf-8'
     });
@@ -93,7 +93,7 @@ export async function getStaticProps() {
   return {
     props: {
       data: data,
-      title: 'Programação'
+      title: 'Outros assuntos legais'
     }
   };
 }
